Small changes and tweaks (Should i maybe just rewrite everything?)
-made some more small changes to make the implementation of better
pathfinding and communation between entities easier

TO DO

-revise the "AI" of dude fixing his priorities
-implement pathfinding

+++ C#/ThinkBot/ThinkBot/Mapping/LocationBlock.cs
@@ -8,11 +8,11 @@ namespace ThinkBot.Mapping
 {
     public class LocationBlock
     {
-        public int PosX { get; set; }
-        public int PosY { get; set; }
+        public short PosX { get; set; }
+        public short PosY { get; set; }
         public bool IsPassable { get; set; }
 
-        public LocationBlock(bool isPassable, int posX, int posY)
+        public LocationBlock(bool isPassable, short posX, short posY)
         {
             IsPassable = isPassable;
             PosX = posX;
